Using InputStreams as much as possible
git-svn-id: https://scm.mmbase.org/mmbase/trunk@36963 91430c5f-d668-0410-9e7d-dd663366dcfd

+++ applications/resources/src/main/java/org/mmbase/util/images/AbstractImageConverter.java
@@ -53,7 +53,7 @@ public abstract class AbstractImageConverter implements ImageConverter {
     /**
      * @see org.mmbase.util.images.ImageConverter#convertImage(java.io.InputStream, java.lang.String, java.io.OutputStream, java.util.List)
      */
-    public int convertImage(InputStream input, String sourceFormat, OutputStream out, List<String> commands) throws IOException {
+    public long convertImage(InputStream input, String sourceFormat, OutputStream out, List<String> commands) throws IOException {
         byte[] bytes;
         if (input instanceof BytesInputStream) {
             bytes = ((BytesInputStream) input).getBuffer();
